// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ConversationType = {
  PRIVATE: 'PRIVATE',
  GROUP: 'GROUP'
} as const

export type ConversationType = (typeof ConversationType)[keyof typeof ConversationType]


export const ProjectStatus = {
  NEW: 'NEW',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED'
} as const

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus]


export const TaskStatus = {
  NEW: 'NEW',
  IN_PROGRESS: 'IN_PROGRESS',
  PAUSED: 'PAUSED',
  TO_CHECK: 'TO_CHECK',
  COMPLETED: 'COMPLETED',
  RETURNED: 'RETURNED'
} as const

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus]


export const TaskPriority = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH'
} as const

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority]


export const UserRole = {
  ROOT: 'ROOT',
  ADMIN: 'ADMIN',
  PROJECT_MANAGER: 'PROJECT_MANAGER',
  DEVELOPER: 'DEVELOPER'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const UserStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  BANNED: 'BANNED',
  DELETED: 'DELETED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const MessageStatus = {
  SENT: 'SENT',
  DELIVERED: 'DELIVERED',
  READ: 'READ'
} as const

export type MessageStatus = (typeof MessageStatus)[keyof typeof MessageStatus]


export const LogAction = {
  CREATE: 'CREATE',
  UPDATE: 'UPDATE',
  DELETE: 'DELETE',
  LOGIN: 'LOGIN',
  LOGOUT: 'LOGOUT'
} as const

export type LogAction = (typeof LogAction)[keyof typeof LogAction]


export const LogEntity = {
  PROJECT: 'PROJECT',
  TASK: 'TASK',
  USER: 'USER',
  CATEGORY: 'CATEGORY'
} as const

export type LogEntity = (typeof LogEntity)[keyof typeof LogEntity]